Conflicts:
	src/lpj/fprintconfig.c
	src/lpj/fscanconfig.c
	src/lpj/update_daily.c

+++ src/utils/cdf2bin.c
@@ -349,7 +349,7 @@ int main(int argc,char **argv)
       return EXIT_FAILURE;
     }
     if(getfilesize(argv[i]) % (sizeof(short)*2)!=0)
-      fprintf(stderr,"Warning: File size of '%s' is not multiple of %d.\n",argv[i],(int)(sizeof(float)*2));
+      fprintf(stderr,"Warning: File size of '%s' is not multiple of %d.\n",argv[i],(int)(sizeof(short)*2));
     config.ngridcell=getfilesize(argv[i])/sizeof(short)/2;
     if(config.ngridcell==0)
     {
